token generation migrated to pkce module
