+++ examples/login_demo.html
@@ -77,7 +77,6 @@
     import { SlashDBClient, DataDiscoveryDatabase, DataDiscoveryResource, DataDiscoveryFilter, SQLPassThruQuery, SQLPassThruFilter,
              eq, any, between, gte, lte, not, and, asc, desc, chgPlaceHolder } from '../src/index.js';
     import { PKCE } from '../src/pkce.js';
-    import { generateCodeVerifier, base64urlencode } from '../src/utils.js';
 
     // make methods, modules, available to console for testing
     window.DataDiscoveryFilter = DataDiscoveryFilter
@@ -98,8 +97,6 @@
     window.chgPlaceHolder = chgPlaceHolder;
     
     window.PKCE = PKCE;
-    window.generateCodeVerifier = generateCodeVerifier;
-    window.base64urlencode = base64urlencode;
 
     var host;
     var database;
