+++ src/slashdbclient.js
@@ -1,8 +1,8 @@
 import { DataDiscoveryDatabase } from './datadiscovery.js'
 import { SQLPassThruQuery } from './sqlpassthru.js'
 import { BaseRequestHandler } from './baserequesthandler.js';
-import { PKCE } from './pkce.js';
-import { generateCodeChallenge, generateCodeVerifier, getUrlParms, isSSOredirect, popupCenter } from "./utils.js";
+import { PKCE, generateCodeVerifier, generateCodeChallenge } from './pkce.js';
+import { getUrlParms, isSSOredirect, popupCenter } from "./utils.js";
 
 const SDB_SDBC_INVALID_HOSTNAME = 'Invalid hostname parameter, must be string';
 const SDB_SDBC_INVALID_USERNAME = 'Invalid username parameter, must be string';
